move over styledtext (#3796)
